Closes #1199.

+++ mongodb/src/test/scala/quasar/physical/mongodb/planner.scala
@@ -329,7 +329,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                  $and(
                    $lte($literal(Bson.Date(Instant.ofEpochMilli(0))), $field("baz")),
                    $lt($field("baz"), $literal(Bson.Regex("", "")))),
-                 $add($dayOfWeek($field("baz")), $literal(Bson.Int64(-1))),
+                 $add($dayOfWeek($field("baz")), $literal(Bson.Int32(-1))),
                  $literal(Bson.Undefined))),
            IgnoreId)))
     }
@@ -345,9 +345,9 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                  $and(
                    $lte($literal(Bson.Date(Instant.ofEpochMilli(0))), $field("baz")),
                    $lt($field("baz"), $literal(Bson.Regex("", "")))),
-                 $cond($eq($dayOfWeek($field("baz")), $literal(Bson.Int64(1))),
-                   $literal(Bson.Int64(7)),
-                   $add($dayOfWeek($field("baz")), $literal(Bson.Int64(-1)))),
+                 $cond($eq($dayOfWeek($field("baz")), $literal(Bson.Int32(1))),
+                   $literal(Bson.Int32(7)),
+                   $add($dayOfWeek($field("baz")), $literal(Bson.Int32(-1)))),
                  $literal(Bson.Undefined))),
            IgnoreId)))
     }
@@ -363,7 +363,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                  $and(
                    $lte($literal(Bson.Date(Instant.ofEpochMilli(0))), $field("ts")),
                    $lt($field("ts"), $literal(Bson.Regex("", "")))),
-                 $eq($year($field("ts")), $literal(Bson.Int64(2016))),
+                 $eq($year($field("ts")), $literal(Bson.Int32(2016))),
                  $literal(Bson.Undefined)),
              "__tmp3" -> $$ROOT),
            IgnoreId),
@@ -446,7 +446,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                 $literal(Bson.Undefined)))),
           \/-($literal(Bson.Null))),
         $project(
-          reshape("0" -> $multiply($field("__tmp4"), $literal(Bson.Int64(100)))),
+          reshape("0" -> $multiply($field("__tmp4"), $literal(Bson.Int32(100)))),
           IgnoreId)))
     }
 
@@ -465,7 +465,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                    $and(
                      $lte($literal(Bson.Bool(false)), $field("pop")),
                      $lt($field("pop"), $literal(Bson.Regex("", ""))))),
-                 $cond($lt($field("pop"), $literal(Bson.Int64(10000))),
+                 $cond($lt($field("pop"), $literal(Bson.Int32(10000))),
                    $field("city"),
                    $field("loc")),
                  $literal(Bson.Undefined))),
@@ -507,7 +507,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                Selector.Doc(BsonField.Name("bar") ->
                  Selector.Type(BsonType.Bool)))),
            Selector.Doc(BsonField.Name("bar") ->
-             Selector.Gt(Bson.Int64(10)))))))
+             Selector.Gt(Bson.Int32(10)))))))
     }
 
     "plan simple reversed filter" in {
@@ -530,7 +530,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                Selector.Doc(BsonField.Name("bar") ->
                  Selector.Type(BsonType.Bool)))),
            Selector.Doc(BsonField.Name("bar") ->
-             Selector.Gt(Bson.Int64(10)))))))
+             Selector.Gt(Bson.Int32(10)))))))
     }
 
     "plan simple filter with expression in projection" in {
@@ -553,7 +553,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                Selector.Doc(BsonField.Name("bar") ->
                  Selector.Type(BsonType.Bool)))),
            Selector.Doc(BsonField.Name("bar") ->
-             Selector.Gt(Bson.Int64(10))))),
+             Selector.Gt(Bson.Int32(10))))),
          $project(
            reshape("0" ->
              $cond(
@@ -660,9 +660,9 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                    Selector.Type(BsonType.Bool)))),
              Selector.And(
                Selector.Doc(BsonField.Name("bar") ->
-                 Selector.Gte(Bson.Int64(10))),
+                 Selector.Gte(Bson.Int32(10))),
                Selector.Doc(BsonField.Name("bar") ->
-                 Selector.Lte(Bson.Int64(100))))))))
+                 Selector.Lte(Bson.Int32(100))))))))
     }
 
     "plan filter with like" in {
@@ -841,7 +841,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                    Selector.Type(BsonType.Bool)))),
              Selector.And(
                Selector.Doc(BsonField.Name("bar") ->
-                 Selector.Neq(Bson.Int64(-10))),
+                 Selector.Neq(Bson.Int32(-10))),
                Selector.Doc(BsonField.Name("baz") ->
                  Selector.Gt(Bson.Dec(-1.0))))))))
     }
@@ -867,7 +867,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                  Selector.Doc(BsonField.Name("bar") ->
                    Selector.Type(BsonType.Bool)))),
              Selector.And(
-               Selector.Doc(BsonField.Name("bar") -> Selector.Gt(Bson.Int64(10))),
+               Selector.Doc(BsonField.Name("bar") -> Selector.Gt(Bson.Int32(10))),
                Selector.Or(
                  Selector.Doc(BsonField.Name("baz") ->
                    Selector.Eq(Bson.Text("quux"))),
@@ -913,9 +913,9 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                      Selector.Type(BsonType.Bool)))),
                Selector.Or(
                  Selector.Doc(ListMap[BsonField, Selector.SelectorExpr](
-                   BsonField.Name("pop") -> Selector.NotExpr(Selector.Gt(Bson.Int64(0))))),
+                   BsonField.Name("pop") -> Selector.NotExpr(Selector.Gt(Bson.Int32(0))))),
                  Selector.Doc(ListMap[BsonField, Selector.SelectorExpr](
-                   BsonField.Name("pop") -> Selector.NotExpr(Selector.Lt(Bson.Int64(1000)))))))))))
+                   BsonField.Name("pop") -> Selector.NotExpr(Selector.Lt(Bson.Int32(1000)))))))))))
     }
 
     "plan filter with not and equality" in {
@@ -924,7 +924,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
           $read(Collection("db", "zips")),
           $match(
             Selector.Doc(ListMap[BsonField, Selector.SelectorExpr](
-              BsonField.Name("pop") -> Selector.NotExpr(Selector.Eq(Bson.Int64(0))))))))
+              BsonField.Name("pop") -> Selector.NotExpr(Selector.Eq(Bson.Int32(0))))))))
     }
 
     "plan filter with \"is not null\"" in {
@@ -982,7 +982,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
             "__tmp0" -> $sum($literal(Bson.Int32(1)))),
           -\/(reshape("0" -> $field("city")))),
         $match(Selector.Doc(
-          BsonField.Name("__tmp0") -> Selector.Gt(Bson.Int64(10)))),
+          BsonField.Name("__tmp0") -> Selector.Gt(Bson.Int32(10)))),
         $project(
           reshape("city" -> $field("_id", "0")),
           ExcludeId)))
@@ -1004,7 +1004,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                   $literal(Bson.Undefined)))),
           -\/(reshape("0" -> $field("city")))),
         $match(Selector.Doc(
-          BsonField.Name("1") -> Selector.Gt(Bson.Int64(50000)))),
+          BsonField.Name("1") -> Selector.Gt(Bson.Int32(50000)))),
         $project(
           reshape(
             "city" -> $field("_id", "0"),
@@ -1046,7 +1046,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                     $lt($field("pop"), $literal(Bson.Regex("", ""))))),
                 $and(
                   $neq($field("city"), $field("state")),
-                  $lt($field("pop"), $literal(Bson.Int64(10000)))),
+                  $lt($field("pop"), $literal(Bson.Int32(10000)))),
                 $literal(Bson.Undefined)),
             "__tmp5" -> $$ROOT),
           IgnoreId),
@@ -1076,8 +1076,8 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                     $lt($field("pop"), $literal(Bson.Text("")))),
                   $substr(
                     $literal(Bson.Text("fghijklmnop")),
-                    $literal(Bson.Int64(0)),
-                    $divide($field("pop"), $literal(Bson.Int64(10000)))),
+                    $literal(Bson.Int32(0)),
+                    $divide($field("pop"), $literal(Bson.Int32(10000)))),
                   $literal(Bson.Undefined))),
             IgnoreId)))
     }
@@ -1141,7 +1141,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                     $and(
                       $lte($literal(Bson.Date(Instant.ofEpochMilli(0))), $field("bar")),
                       $lt($field("bar"), $literal(Bson.Regex("", ""))))),
-                  $divide($field("bar"), $literal(Bson.Int64(10))),
+                  $divide($field("bar"), $literal(Bson.Int32(10))),
                   $literal(Bson.Undefined))),
             IgnoreId),
           $sort(NonEmptyList(BsonField.Name("__tmp2") -> Ascending)),
@@ -1213,7 +1213,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                 Selector.Doc(BsonField.Name("pop") ->
                   Selector.Type(BsonType.Bool)))),
             Selector.Doc(
-              BsonField.Name("pop") -> Selector.Lt(Bson.Int64(1000))))),
+              BsonField.Name("pop") -> Selector.Lt(Bson.Int32(1000))))),
           $simpleMap(
             NonEmptyList(MapExpr(JsFn(Name("x"),
               obj(
@@ -1296,7 +1296,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                     $and(
                       $lte($literal(Bson.Date(Instant.ofEpochMilli(0))), $field("pop")),
                       $lt($field("pop"), $literal(Bson.Regex("", ""))))),
-                  $divide($field("pop"), $literal(Bson.Int64(1000))),
+                  $divide($field("pop"), $literal(Bson.Int32(1000))),
                   $literal(Bson.Undefined))),
             IgnoreId),
           $sort(NonEmptyList(BsonField.Name("popInK") -> Ascending))))
@@ -1322,7 +1322,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                 Selector.Doc(BsonField.Name("pop") ->
                   Selector.Type(BsonType.Bool)))),
             Selector.Doc(
-              BsonField.Name("pop") -> Selector.Lte(Bson.Int64(1000))))),
+              BsonField.Name("pop") -> Selector.Lte(Bson.Int32(1000))))),
           $project(
             reshape(
               "city" -> $field("city"),
@@ -1353,7 +1353,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                 Selector.Doc(BsonField.Name("pop") ->
                   Selector.Type(BsonType.Bool)))),
 
-            Selector.Doc(BsonField.Name("pop") -> Selector.Gte(Bson.Int64(1000))))),
+            Selector.Doc(BsonField.Name("pop") -> Selector.Gte(Bson.Int32(1000))))),
           $project(
             reshape(
               "popInK" ->
@@ -1365,7 +1365,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                     $and(
                       $lte($literal(Bson.Date(Instant.ofEpochMilli(0))), $field("pop")),
                       $lt($field("pop"), $literal(Bson.Regex("", ""))))),
-                  $divide($field("pop"), $literal(Bson.Int64(1000))),
+                  $divide($field("pop"), $literal(Bson.Int32(1000))),
                   $literal(Bson.Undefined))),
             ExcludeId),
           $sort(NonEmptyList(BsonField.Name("popInK") -> Ascending)))
@@ -1691,7 +1691,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
             $unwind(DocField("pop")),
             $project(
               reshape(
-                "0"   -> $divide($field("__tmp3"), $literal(Bson.Int64(1000))),
+                "0"   -> $divide($field("__tmp3"), $literal(Bson.Int32(1000))),
                 "pop" -> $field("pop")),
               IgnoreId))
         }
@@ -1766,7 +1766,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                     $and(
                       $lte($literal(Bson.Date(Instant.ofEpochMilli(0))), $field("pop")),
                       $lt($field("pop"), $literal(Bson.Regex("", ""))))),
-                  $divide($field("pop"), $literal(Bson.Int64(1000))),
+                  $divide($field("pop"), $literal(Bson.Int32(1000))),
                   $literal(Bson.Undefined))),
             "__tmp6" -> reshape(
               "__tmp2" ->
@@ -1803,13 +1803,13 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                   $and(
                     $lt($literal(Bson.Null), $field("pop")),
                     $lt($field("pop"), $literal(Bson.Text("")))),
-                  $subtract($field("pop"), $literal(Bson.Int64(1))),
+                  $subtract($field("pop"), $literal(Bson.Int32(1))),
                   $literal(Bson.Undefined)))),
           -\/(reshape("0" -> $field("city")))),
         $project(
           reshape(
             "city" -> $field("_id", "0"),
-            "1"    -> $divide($field("__tmp6"), $literal(Bson.Int64(1000)))),
+            "1"    -> $divide($field("__tmp6"), $literal(Bson.Int32(1000)))),
           IgnoreId)))
     }
 
@@ -2084,7 +2084,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
           IgnoreId),
         $unwind(DocField(BsonField.Name("__tmp5"))),
         $match(Selector.Doc(
-          BsonField.Name("__tmp5") -> Selector.Lt(Bson.Int64(0)))),
+          BsonField.Name("__tmp5") -> Selector.Lt(Bson.Int32(0)))),
         $project(
           reshape("loc" -> $field("__tmp5")),
           IgnoreId)))
@@ -2114,7 +2114,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                 $and(
                   $lte($literal(Bson.Text("")), $field("__tmp12", "sha")),
                   $lt($field("__tmp12", "sha"), $literal(Bson.Doc(ListMap())))),
-                $substr($field("__tmp12", "sha"), $literal(Bson.Int64(0)), $literal(Bson.Int64(1))),
+                $substr($field("__tmp12", "sha"), $literal(Bson.Int32(0)), $literal(Bson.Int32(1))),
                 $literal(Bson.Undefined))))),
         $project(
           reshape(
@@ -2332,8 +2332,8 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                     $lt($field("city"), $literal(Bson.Doc(ListMap())))),
                   $substr(
                     $field("city"),
-                    $literal(Bson.Int64(0)),
-                    $literal(Bson.Int64(1))),
+                    $literal(Bson.Int32(0)),
+                    $literal(Bson.Int32(1))),
                   $literal(Bson.Undefined))))),
           $group(
             grouped("0" -> $sum($literal(Bson.Int32(1)))),
@@ -2431,7 +2431,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
           $group(
             grouped(
               "state" -> $first($field("__tmp0", "state")),
-              "1"     -> $sum($literal(Bson.Int64(1)))),
+              "1"     -> $sum($literal(Bson.Int32(1)))),
             \/-($literal(Bson.Null)))))
     }.pendingUntilFixed
 
@@ -2533,7 +2533,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
                 Selector.Doc(BsonField.Name("pop") ->
                   Selector.Type(BsonType.Bool)))),
             Selector.Doc(
-              BsonField.Name("pop") -> Selector.Gt(Bson.Int64(1000))))),
+              BsonField.Name("pop") -> Selector.Gt(Bson.Int32(1000))))),
           $simpleMap(NonEmptyList(MapExpr(JsFn(Name("x"), obj(
             "city"   -> Select(ident("x"), "city"),
             "pop"    -> Select(ident("x"), "pop"),
@@ -3642,7 +3642,7 @@ class PlannerSpec extends Specification with ScalaCheck with CompilerHelpers wit
       plan(lp) must beWorkflow(chain(
         $read(Collection("db", "foo")),
         $match(Selector.Doc(
-          BsonField.Name("baz") -> Selector.Eq(Bson.Int64(0)))),
+          BsonField.Name("baz") -> Selector.Eq(Bson.Int32(0)))),
         $sort(NonEmptyList(BsonField.Name("bar") -> Ascending))))
     }
